Extract ResizeAttrBase to be reused in TensorResizeAttr
